CI: Use docker image to speed up jobs
Style check job installs dependencies on each run. This
causes that the pipelines take a long time. Let's use a custom docker
image to not block runners for such a long time and to get the pipeline
results faster. This commit also adds a new job to generate the docker
image for us.